brezee_paa/requirements.txt
fastapi==0.104.1 
uvicorn==0.24.0
python-dotenv==1.0.0
pandas==2.1.3

requests==2.31.0
pyppeteer==1.0.2    
pyotp==2.9.0
fyers-apiv3==3.0.5
dhanhq==1.2.2

python-multipart==0.0.6
sqlalchemy==2.0.23

nest_asyncio==1.5.8